Added OPT_PALETTE, OPT_BRIGHTNESS, etc.

+++ GUI/Settings/Defaults.swift
@@ -1344,9 +1344,9 @@ struct VideoDefaults {
     
     // Colors
     let palette: Palette
-    let brightness: Double
-    let contrast: Double
-    let saturation: Double
+    let brightness: Int
+    let contrast: Int
+    let saturation: Int
     
     // Geometry
     let hCenter: Float
@@ -1384,9 +1384,9 @@ struct VideoDefaults {
     static let tft = VideoDefaults.init(
         
         palette: Palette.COLOR,
-        brightness: 50.0,
-        contrast: 100.0,
-        saturation: 50.0,
+        brightness: 50,
+        contrast: 100,
+        saturation: 50,
         
         hCenter: 0.6333,
         vCenter: 0.1683,
@@ -1418,9 +1418,9 @@ struct VideoDefaults {
     static let crt = VideoDefaults.init(
         
         palette: Palette.COLOR,
-        brightness: 50.0,
-        contrast: 100.0,
-        saturation: 50.0,
+        brightness: 50,
+        contrast: 100,
+        saturation: 50,
         
         hCenter: 0.1169,
         vCenter: 0.1683,
